main file rewrite complete
